Refactor http request checksum interceptor
