+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/AwsCodegenDecorator.kt
@@ -44,6 +44,7 @@ val DECORATORS: List<ClientCodegenDecorator> =
             UserAgentDecorator(),
             SigV4AuthDecorator(),
             HttpRequestChecksumDecorator(),
+            HttpRequestChecksumMutationInterceptorGenerator(),
             HttpResponseChecksumDecorator(),
             IntegrationTestDecorator(),
             AwsFluentClientDecorator(),
