+++ codegen-client/src/main/kotlin/software/amazon/smithy/rust/codegen/client/smithy/generators/OperationGenerator.kt
@@ -217,7 +217,5 @@ open class OperationGenerator(
             .render(operationWriter, operationShape)
 
         HttpResponseChecksumMutationInterceptorGenerator(codegenContext).render(operationWriter, operationShape)
-
-        HttpRequestChecksumMutationInterceptorGenerator(codegenContext).render(operationWriter, operationShape)
     }
 }
